// src/app/shared/models/cig-request.dto.ts
export interface CigRequestDto {
    industrialConsumerId: string;
    industrialConsumerBranchId: string;

    pickupId: string;
    pickupBranchId: string;

    pretreatmentId: string;
    pretreatmentBranchId: string;

    valuerId: string;
    valuerBranchId: string;

    // industrialConsumerId: string;
    // industrialConsumerBranchId: string;

    // pickupId: string;
    // pickupBranchId: string;

    // pretreatmentId: string;
    // pretreatmentBranchId: string;

    // valuerId: string;
    // valuerBranchId: string;

    // status: CigStatus;
}